use default urls and add functional dropbox dialog

+++ app/controllers/services_controller.rb
@@ -18,11 +18,10 @@ class ServicesController < ApplicationController
   # @option params [String] :id the service to use
   def new
     if params[:id] == 'Dropbox'
-      current_user.dropbox_connection.destroy
+      current_user.dropbox_connection.destroy unless current_user.dropbox_connection.nil?
       session = DropboxSession.new ENV['DROPBOX_APP_KEY'], ENV['DROPBOX_APP_SECRET']
       session.get_request_token
-      redirect_to session.get_authorize_url url_for(controller: :services, action: :confirm, id: 'Dropbox',
-                                                    only_path: false, host: 'nimbus-web.herokuapp.com', protocol: 'https')
+      redirect_to session.get_authorize_url url_for(controller: :services, action: :confirm, id: 'Dropbox', only_path: false)
       current_user.create_dropbox_connection session: session.serialize
     end
   end
